first draft of SAX parser implementation
